fix: allow use in bundlers
We can't use the `**` operator as right now it gets transpiled to `Math.pow`
which only takes Numbers.

Similarly, trying to avoid referencing node globals to get a `Buffer` instance
causes similar problems now that bundlers don't tend to come with node polyfills
any more.

Swaps the use of `**` for the BigInt literal result of the calculation and uses
the `Buffer` module to ensure compatibility.

Fixes #7
Fixes #8

Please publish with a `main` field to also fix #9.